refactor: Update BufferedByteReader to always update currBuffLen
This commit updates the BufferedByteReader in buffer.go to always update the currBuffLen variable, regardless of the number of bytes read from the file. This ensures that the currBuffLen accurately reflects the number of bytes in the current buffer. This change improves the reliability and consistency of the BufferedByteReader implementation.